Refactor out another function.
